backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.0
pydantic==1.10.12
reportlab==4.1.3
google-generativeai==0.1.1  
python-multipart==0.0.6     